Rename cmdline driver package
rename package to a less generic name.
this is needed for clarity when additional
TC drivers are introduced.
