+++ pkg/tc/driver/cmdline/converters.go
@@ -1,4 +1,4 @@
-package driver
+package cmdline
 
 import (
 	"strings"
